mas: better handle signin failure
- see if app can be skipped because it doesn’t need upgrade
- don’t try to sign in on Mojave (it doesn’t work)
- only fail if app actually needs installed/upgraded
- don’t print `mas account` output; just use exit code
- add `MacOS` stub and cleanup existing stub code
