+++ spec/mac_app_store_installer_spec.rb
@@ -62,7 +62,7 @@ describe Bundle::MacAppStoreInstaller do
       end
 
       it "tries to sign in with mas" do
-        expect(Bundle).to receive(:system).with("mas", "account").and_return(false).twice
+        expect(Kernel).to receive(:system).with("mas account &>/dev/null").and_return(false).twice
         expect(Bundle).to receive(:system).with("mas", "signin", "--dialog", "").and_return(true)
         expect { do_install }.to raise_error(RuntimeError)
       end
