+++ spec/stub/formulary.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class Formulary
+  def self.factory(name)
+    Formula.new(name)
+  end
+end
