+++ spec/stub/global.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+HOMEBREW_PREFIX = Pathname.new("/usr/local")
+HOMEBREW_REPOSITORY = Pathname.new("/usr/local/Homebrew")
