+++ spec/stub/os.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module MacOS
+  module_function
+
+  def version
+    :high_sierra
+  end
+end
