+++ spec/stub/tap.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class Tap
+  def self.map
+    []
+  end
+end
